requirements.txt
discord.py>=2.0.0
java
wavelink
python-dotenv>=0.19.2
flask-sslify>=0.1
flask>=2.0.2